+++ King.py
@@ -1,0 +1,5 @@
+from piece import Piece
+
+class King(Piece):
+    def __init__(self, color, position=None):
+        super().__init__(color, position)